requirements.txt
flask
google-generativeai
python-pptx
gTTS
requests
python-dotenv
beautifulsoup4